move blue card sorting list to components
